+++ c_station/service/service_video/src/main/java/cn/o0u0o/service/video/service/VVideoExtraService.java
@@ -14,5 +14,5 @@ import com.baomidou.mybatisplus.extension.service.IService;
  */
 public interface VVideoExtraService extends IService<VVideoExtra> {
 
-    boolean addVideo(Long videoId, VideoUpload video);
+    boolean addVideo(Integer videoId, VideoUpload video);
 }
