+++ c_station/service/service_video/src/main/java/cn/o0u0o/service/video/service/VVideoItemService.java
@@ -14,5 +14,5 @@ import com.baomidou.mybatisplus.extension.service.IService;
  */
 public interface VVideoItemService extends IService<VVideoItem> {
 
-    boolean addVideo(Long videoId, Long size, VideoUpload video);
+    boolean addVideo(Integer videoId, Long size, VideoUpload video);
 }
